requirements.txt
# FastAPI y dependencias
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Utilidades
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Authentication
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
